........
  r75220 | mark.dickinson | 2009-10-03 21:13:37 +0100 (Sat, 03 Oct 2009) | 1 line

  Use size_t instead of int for a PyMem_MALLOC argument;  silences a gcc 'comparison always false' warning.
........

+++ Objects/capsule.c
@@ -197,7 +197,7 @@ PyCapsule_Import(const char *name, int no_block)
     PyObject *object = NULL;
     void *return_value = NULL;
     char *trace;
-    int name_length = (strlen(name) + 1) * sizeof(char);
+    size_t name_length = (strlen(name) + 1) * sizeof(char);
     char *name_dup = (char *)PyMem_MALLOC(name_length);
 
     if (!name_dup) {
